Modify DAO due to database change

+++ src/main/java/controller/DAO/ContactDAO.java
@@ -49,7 +49,7 @@ public class ContactDAO {
         ArrayList<String> phone = cont.getPhonesList();
         ArrayList<String> mail = cont.getMailsList();
         boolean reserved = cont.isReserved();
-        String role = cont.getRole();
+        int role = RoleDAO.getInstance().getRoleByName(cont.getRole());
         Entity entity = cont.getEntity();
         String entityname = null;
         int referent=-1;
@@ -66,14 +66,16 @@ public class ContactDAO {
         String req_add ="INSERT INTO Contact(name,surname,address,role,entity) VALUES (?,?,?,?,?)";
        
         try {
-            
+            if (role==-1){
+                throw new Exception("Invalid Role: "+cont.getRole());
+            }
             PreparedStatement req_add_prep =null;
             if (reserved){
                 req_add_prep = this.db.prepareStatement(req_add_reserved,Statement.RETURN_GENERATED_KEYS);
                 req_add_prep.setString(1,name);
                 req_add_prep.setString(2,surname);
                 req_add_prep.setString(3,address);
-                req_add_prep.setString(4,role);
+                req_add_prep.setInt(4,role);
                 req_add_prep.setInt(5,referent);
                 req_add_prep.setString(6,entityname);
             }
@@ -82,7 +84,7 @@ public class ContactDAO {
                 req_add_prep.setString(1,name);
                 req_add_prep.setString(2,surname);
                 req_add_prep.setString(3,address);
-                req_add_prep.setString(4,role);
+                req_add_prep.setInt(4,role);
                 req_add_prep.setString(5,entityname);
             }
             int insert= req_add_prep.executeUpdate();
@@ -183,7 +185,7 @@ public class ContactDAO {
         ArrayList<String> phone = cont.getPhonesList();
         ArrayList<String> mail = cont.getMailsList();
         boolean reserved= cont.isReserved();
-        String role = cont.getRole();
+        int role = RoleDAO.getInstance().getRoleByName(cont.getRole());
         Entity entity = cont.getEntity();
         String entityname = null;
         int referent=-1;
@@ -198,6 +200,9 @@ public class ContactDAO {
         String req_update_ref = "UPDATE Contact set name= ?,surname= ?,address= ?,role= ?, referent=?,entity=? where id=?";
         
         try {
+            if (role==-1){
+                throw new Exception("Invalid Role: "+cont.getRole());
+            }
             // --Update associated mails and phones
             this.deleteMail(id);
             this.deletePhone(id);
@@ -211,7 +216,7 @@ public class ContactDAO {
                 req_update_prep.setString(1,name);
                 req_update_prep.setString(2,surname);
                 req_update_prep.setString(3,address);
-                req_update_prep.setString(4,role);
+                req_update_prep.setInt(4,role);
                 req_update_prep.setInt(5,referent);
                 req_update_prep.setString(6,entityname);
                  //WHERE
@@ -222,7 +227,7 @@ public class ContactDAO {
                 req_update_prep.setString(1,name);
                 req_update_prep.setString(2,surname);
                 req_update_prep.setString(3,address);
-                req_update_prep.setString(4,role);
+                req_update_prep.setInt(4,role);
                 req_update_prep.setString(5,entityname);
                 //WHERE
                 req_update_prep.setInt(6, id);
@@ -344,7 +349,7 @@ public class ContactDAO {
         assert acc.getSectors()!=null : "Pre condition violated";
         assert acc.getSectors().size()>0 && !acc.getContact().isLinkAccount(): "Pre condition violated";
 
-        return AccountDAO.getInstance().getAccountByName(acc.getUsername()).getContact();
+        return AccountDAO.getInstance().getAccountById(acc.getId()).getContact();
     }
 
 
@@ -358,21 +363,21 @@ public class ContactDAO {
         assert sec!=null && sec.getName()!=null && !sec.getName().isEmpty() :"Pre condition violated";
         //Data
         int id;
-        String nameSec = sec.getName();
+        int sec_id = sec.getId(); 
         String name;
         String surname;
         String address;
         boolean reserved;
         String role;
         int refId;
-        String entity;
+        int entity;
         //Request
-        String req_select = "SELECT cont.id,cont.name,cont.surname,cont.address,cont.reserved,cont.referent,cont.role,cont.entity FROM Contact_Sector_Asso AS sec_con,Contact AS cont WHERE sec_con.contactId=cont.id AND sectorName=? ORDER BY cont.name,cont.surname ASC";
+        String req_select = "SELECT cont.id,cont.name,cont.surname,cont.address,cont.reserved,cont.referent,cont.role,cont.entity FROM Contact_Sector_Asso AS sec_con,Contact AS cont WHERE sec_con.contactId=cont.id AND sectorId=? ORDER BY cont.name,cont.surname ASC";
         ArrayList<Contact> ret = new ArrayList<>();
         try {
 
             PreparedStatement req_select_prep = this.db.prepareStatement(req_select);
-            req_select_prep.setString(1,nameSec );
+            req_select_prep.setInt(1,sec_id );
             ResultSet res = req_select_prep.executeQuery();
             while (res.next()) {
                 id = res.getInt("cont.id");
@@ -382,7 +387,7 @@ public class ContactDAO {
                 reserved = res.getBoolean("cont.reserved");
                 role = res.getString("cont.role");
                 refId = res.getInt("cont.referent");
-                entity = res.getString("cont.entity");
+                entity = res.getInt("cont.entity");
                
                 Contact ref = null;
                 if (reserved){
@@ -391,7 +396,7 @@ public class ContactDAO {
                 Contact ctc = new Contact(name, surname,role,ref,reserved,id);
                 this.getMailByContact(ctc);
                 this.getPhoneByContact(ctc);
-                ctc.setEntity(EntityDAO.getInstance().getEntityByName(entity));
+                ctc.setEntity(EntityDAO.getInstance().getEntityById(entity));
                 ctc.setAddress(address);
                 ret.add(ctc);
             }
@@ -423,7 +428,7 @@ public class ContactDAO {
         boolean reserved;
         String role;
         int refId;
-        String entity;
+        int entity;
         //Request
         String req_select = "SELECT cont.id,cont.name,cont.surname,cont.address,cont.reserved,cont.referent,cont.role,cont.entity FROM Contact_Event_Asso AS evt_con,Contact AS cont WHERE evt_con.contactId=cont.id AND eventId=? ORDER BY cont.name,cont.surname ASC";
         ArrayList<Contact> ret = new ArrayList<>();
@@ -441,7 +446,7 @@ public class ContactDAO {
                 reserved = res.getBoolean("cont.reserved");
                 role = res.getString("cont.role");
                 refId = res.getInt("cont.referent");
-                entity = res.getString("cont.entity");
+                entity = res.getInt("cont.entity");
     
                
                 Contact ref = null;
@@ -453,7 +458,7 @@ public class ContactDAO {
                 this.getMailByContact(ctc);
                 this.getPhoneByContact(ctc);
                 ctc.setAddress(address);
-                ctc.setEntity(EntityDAO.getInstance().getEntityByName(entity));
+                ctc.setEntity(EntityDAO.getInstance().getEntityById(entity));
                 ret.add(ctc);
             }
 
@@ -481,7 +486,7 @@ public class ContactDAO {
         String address;
         boolean reserved;
         String role;
-        String entity;
+        int entity;
         int refId;
         //Request
         try {
@@ -497,13 +502,13 @@ public class ContactDAO {
                 reserved = res.getBoolean("reserved");
                 role = res.getString("role");
                 refId = res.getInt("referent");
-                entity= res.getString("entity");
+                entity= res.getInt("entity");
                 Contact ref = null;
                 if (reserved){
                     ref =this.getContactById(refId);
                 }
                 Contact contact=  new Contact(name,surname,role,ref,reserved,id);
-                contact.setEntity(EntityDAO.getInstance().getEntityByName(entity));
+                contact.setEntity(EntityDAO.getInstance().getEntityById(entity));
                 contact.setAddress(address);
                 this.getMailByContact(contact);
                 this.getPhoneByContact(contact);
@@ -535,7 +540,7 @@ public class ContactDAO {
         boolean reserved;
         String role;
         int refId;
-        String entity;
+        int entity;
         //Request
         try {
 
@@ -550,13 +555,13 @@ public class ContactDAO {
                 reserved = res.getBoolean("reserved");
                 role = res.getString("role");
                 refId = res.getInt("referent");
-                entity = res.getString("entity");
+                entity = res.getInt("entity");
                 Contact ref = null;
                 if (reserved){
                     ref =this.getContactById(refId);
                 }
                 Contact contact=  new Contact(name,surname,role,ref,reserved,id);
-                contact.setEntity(EntityDAO.getInstance().getEntityByName(entity));
+                contact.setEntity(EntityDAO.getInstance().getEntityById(entity));
                 contact.setAddress(address);
                 this.getMailByContact(contact);
                 this.getPhoneByContact(contact);
@@ -584,7 +589,7 @@ public class ContactDAO {
         boolean reserved;
         String role;
         int refId;
-        String entity;
+        int entity;
         Contact ref;
         //Request
         String req_select_all= "SELECT * FROM Contact ORDER BY name,surname ASC";
@@ -601,7 +606,7 @@ public class ContactDAO {
                 reserved = res.getBoolean("reserved");
                 role = res.getString("role");
                 refId = res.getInt("referent");
-                entity = res.getString("entity");
+                entity = res.getInt("entity");
                 ref = null;
                 if (reserved){
                     ref =this.getContactById(refId);
@@ -611,7 +616,7 @@ public class ContactDAO {
                 contact.setAddress(address);
                 this.getMailByContact(contact);
                 this.getPhoneByContact(contact);
-                contact.setEntity(EntityDAO.getInstance().getEntityByName(entity));
+                contact.setEntity(EntityDAO.getInstance().getEntityById(entity));
                 ret.add(contact);
             }
 
@@ -677,7 +682,7 @@ public class ContactDAO {
         String surname;
         String address;
         boolean reserved;
-        String entity;
+        int entity;
         String role;
         Contact ref;
         int refId;
@@ -694,14 +699,14 @@ public class ContactDAO {
                 reserved = res.getBoolean("reserved");
                 role = res.getString("role");
                 refId = res.getInt("referent");
-                entity = res.getString("entity");
+                entity = res.getInt("entity");
                 Contact referent = null;
                 if (reserved){
                     referent=this.getContactById(refId);
                 }
                 Contact contact=  new Contact(name,surname,role,referent,reserved,id);
                 contact.setAddress(address);
-                contact.setEntity(EntityDAO.getInstance().getEntityByName(entity));
+                contact.setEntity(EntityDAO.getInstance().getEntityById(entity));
                 this.getMailByContact(contact);
                 this.getPhoneByContact(contact);
                 ret=contact;
@@ -717,21 +722,7 @@ public class ContactDAO {
      * @return list of roles
      */
     public ArrayList<String> getAllRoles(){
-        ArrayList<String> ret = new ArrayList<>();
-        String role;
-        String req_select_roles = "SELECT * FROM Contact_role ORDER BY name ASC";
-        try {
-            PreparedStatement  req_select_roles_prep = this.db.prepareStatement( req_select_roles);
-            ResultSet res = req_select_roles_prep.executeQuery();
-            while (res.next()){
-                role = res.getString("name");
-                ret.add(role);
-            }
-        }
-        catch (Exception e){
-            e.printStackTrace();
-        }
-        return ret;
+        return RoleDAO.getInstance().getAllRoles();
     }
